Removed PairingResponseStatus and created DeRecPairingStatus (#24 and #26 fix)
