+++ src/main/java/org/derecalliance/derec/api/DeRecHelperStatus.java
@@ -20,19 +20,6 @@ package org.derecalliance.derec.api;
 /**
  * Representation of a helper as perceived by a sharer
  */
-public interface DeRecHelperStatus {
+public interface DeRecHelperStatus extends DeRecPairingStatus {
     DeRecIdentity getId();
-
-    PairingStatus getStatus();
-
-    enum PairingStatus {
-        NONE, // not yet invited
-        INVITED, // no reply yet
-        PAIRED, // replied positively
-        REFUSED, // replied negatively
-        PENDING_REMOVAL, // in the process of being removed
-        REMOVED, // at sharer request
-        FAILED, // timeout, disconnect etc.
-        GONE // disconnected at Helper Request
-    }
 }
